add main.py
