+++ sketch-rnn/readme.txt
@@ -1,0 +1,3 @@
+run python3 main.py, which will call get_sketch_dictionary inside sketch_output.py.
+
+The return value of get_sketch_dictionary is a dictionary of lines which is what our application backend desires.
